+++ benchmark/build.gradle
@@ -6,4 +6,7 @@ dependencies {
     compile project(':collekt-vavr')
     compile project(':collekt-pcollections')
     compile project(':collekt-stdlib-jvm')
+
+    compile 'org.openjdk.jmh:jmh-core:1.14.1'
+    compile 'org.openjdk.jmh:jmh-generator-annprocess:1.14.1'
 }
