+++ benchmark/src/main/kotlin/com/github/jcornaz/collekt/benchmark/AddElementBenchmark.kt
@@ -1,4 +1,4 @@
-package com.github.jcornaz.collekt.api
+package com.github.jcornaz.collekt.benchmark
 
 import org.openjdk.jmh.annotations.Benchmark
 import org.openjdk.jmh.annotations.Fork
