+++ benchmark/src/main/kotlin/com/github/jcornaz/collekt/benchmark/IterationBenchmark.kt
@@ -1,21 +1,19 @@
-package com.github.jcornaz.collekt.api
+package com.github.jcornaz.collekt.benchmark
 
 import org.openjdk.jmh.annotations.Benchmark
 import org.openjdk.jmh.annotations.Fork
 import org.openjdk.jmh.annotations.Measurement
 import org.openjdk.jmh.annotations.Warmup
+import org.openjdk.jmh.infra.Blackhole
 import java.util.concurrent.TimeUnit
 
 @Fork(2)
 @Warmup(time = 2, timeUnit = TimeUnit.SECONDS, iterations = 3)
 @Measurement(time = 2, timeUnit = TimeUnit.SECONDS, iterations = 10)
-open class BuildBenchmark {
+open class IterationBenchmark {
 
     @Benchmark
-    fun buildArrayList(input: ArrayListInput) =
-            ArrayList<Int>(input.size.toInt()).apply { addAll(input.size) }
-
-    @Benchmark
-    fun buildImmutableList(input: PersistentInput) =
-            input.factory.from(input.size)
+    fun iteration(input: PersistentInput, blackhole: Blackhole) {
+        input.list.forEach { blackhole.consume(it) }
+    }
 }
