+++ gradle/jvm8.gradle
@@ -1,0 +1,7 @@
+apply from: rootProject.file('gradle/jvm.gradle')
+
+compileKotlin {
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+}
