+++ impl/vavr/src/main/kotlin/com.github.jcornaz.collekt.vavr/VavrList.kt
@@ -1,4 +1,4 @@
-package com.github.jcornaz.collekt.api.impl
+package com.github.jcornaz.collekt.vavr
 
 import com.github.jcornaz.collekt.api.PersistentList
 import com.github.jcornaz.collekt.api.PersistentListFactory
